% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sci_en}
\alias{sci_en}
\title{A sample of English text from PAN BPPT parallel corpora project}
\format{
A vector of character with 6355 elements/lines of sentences
}
\usage{
sci_en
}
\description{
A sample of English text from PAN BPPT parallel corpora project
}
\keyword{datasets}
