% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sci_id}
\alias{sci_id}
\title{A sample of Indonesian text from PAN BPPT parallel corpora project}
\format{
A vector of character with 6355 elements/lines of sentences
}
\usage{
sci_id
}
\description{
A sample of Indonesian text from PAN BPPT parallel corpora project
}
\keyword{datasets}
