% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_conc.R
\name{para_conc}
\alias{para_conc}
\title{Parallel concordance}
\usage{
para_conc(
  source_text = "The source text corpus",
  target_text = "The target text corpus",
  pattern = "Search pattern for words in the source text",
  case_insensitive = FALSE,
  conc_sample = 25,
  filename = "parallel_conc.txt"
)
}
\arguments{
\item{source_text}{character vector of the source-text corpora}

\item{target_text}{character vector of the target-text corpora}

\item{pattern}{regular expression search pattern for the source-text node word}

\item{case_insensitive}{logical; whether the search pattern is case insensitive. Default to FALSE}

\item{conc_sample}{random sample of the concordance lines}

\item{filename}{file name of the parallel concordance output}
}
\value{
A tibble of parallel concordance with source-text node word and its left and right context, and their target-text translation
}
\description{
Parallel concordance
}
\examples{
para_conc(sci_en, sci_id, pattern = "should", conc_sample = 20)
          # we delete the output file to remove warning in R CMD check
          unlink("parallel_conc.txt")
}
